Update ZoneScouter to `v1.6.0`.
* Fixed for the `v0.219.10` PTB patch.
* Bumped up `<LangVersion>` to C# 12.
* Code clean-up and refactoring.
