+++ ZoneScouter/ComfyLib/Extensions/UnityExtensions.cs
@@ -1,7 +1,19 @@
-﻿namespace ComfyLib {
-  public static class ObjectExtensions {
-    public static T Ref<T>(this T unityObject) where T : UnityEngine.Object {
-      return unityObject ? unityObject : null;
+﻿namespace ComfyLib;
+
+using System;
+
+public static class ObjectExtensions {
+  public static T FirstByNameOrThrow<T>(this T[] unityObjects, string name) where T : UnityEngine.Object {
+    foreach (T unityObject in unityObjects) {
+      if (unityObject.name == name) {
+        return unityObject;
+      }
     }
+
+    throw new InvalidOperationException($"Could not find Unity object of type {typeof(T)} with name: {name}");
+  }
+
+  public static T Ref<T>(this T unityObject) where T : UnityEngine.Object {
+    return unityObject ? unityObject : null;
   }
 }
