+++ ZoneScouter/Patches/HudPatch.cs
@@ -1,26 +1,25 @@
-﻿using HarmonyLib;
+﻿namespace ZoneScouter;
 
-using static ZoneScouter.PluginConfig;
-using static ZoneScouter.ZoneScouter;
+using HarmonyLib;
 
-namespace ZoneScouter {
-  [HarmonyPatch(typeof(Hud))]
-  static class HudPatch {
-    [HarmonyPostfix]
-    [HarmonyPatch(nameof(Hud.Awake))]
-    static void AwakePostfix() {
-      if (IsModEnabled.Value) {
-        ToggleSectorInfoPanel();
-        SectorBoundaries.ToggleSectorBoundaries();
-      }
+using static PluginConfig;
+
+[HarmonyPatch(typeof(Hud))]
+static class HudPatch {
+  [HarmonyPostfix]
+  [HarmonyPatch(nameof(Hud.Awake))]
+  static void AwakePostfix() {
+    if (IsModEnabled.Value) {
+      SectorInfoPanelController.ToggleSectorInfoPanel();
+      SectorBoundaries.ToggleSectorBoundaries();
     }
+  }
 
-    [HarmonyPostfix]
-    [HarmonyPatch(nameof(Hud.Update))]
-    static void UpdatePostfix() {
-      if (IsModEnabled.Value && ToggleSectorBoundariesShortcut.Value.IsDown()) {
-        ShowSectorBoundaries.Value = !ShowSectorBoundaries.Value;
-      }
+  [HarmonyPostfix]
+  [HarmonyPatch(nameof(Hud.Update))]
+  static void UpdatePostfix() {
+    if (IsModEnabled.Value && ToggleSectorBoundariesShortcut.Value.IsDown()) {
+      ShowSectorBoundaries.Value = !ShowSectorBoundaries.Value;
     }
   }
 }
