+++ ZoneScouter/manifest.json
@@ -1,6 +1,6 @@
 {
   "name": "ZoneScouter",
-  "version_number": "1.5.0",
+  "version_number": "1.6.0",
   "website_url": "https://github.com/redseiko/ComfyMods/tree/main/ZoneScouter",
   "author": "ComfyMods",
   "description": "The sector's instance level... it's over 9000!",
